refactor: move logic to SummaryContainer, simplify Summary component
